+++ archsim/inc/abi/devices/riscv/SifiveUART.h
@@ -29,8 +29,8 @@ namespace archsim
 					virtual ~SifiveUART();
 
 					bool Initialise() override;
-					bool Read(uint32_t offset, uint8_t size, uint32_t& data) override;
-					bool Write(uint32_t offset, uint8_t size, uint32_t data) override;
+					bool Read(uint32_t offset, uint8_t size, uint64_t& data) override;
+					bool Write(uint32_t offset, uint8_t size, uint64_t data) override;
 
 					bool EnqueueChar(char c);
